:package: refactoring

+++ src/llm/model.rs
@@ -8,7 +8,7 @@ pub struct DocAddingReq {
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct DocAddingSuccess {
     pub id: String,
-    pub embedded: Vec<f64>,
+    pub embedded: Vec<f32>,
 }
 
 #[derive(Debug, Serialize, Deserialize)]
